Integrate rpc-server and metrics-server into new client, remove old client
