+++ Cargo.toml
@@ -32,11 +32,9 @@ members = [
     "consensus",
     "network",
     "client",
-    "client2",
     "rpc-server",
     "metrics-server",
     "lib",
-    "lib2",
     "messages",
     "fixed-unsigned",
     "bls",
