+++ client/Cargo.toml
@@ -18,45 +18,18 @@ is-it-maintained-open-issues = { repository = "nimiq/core-rs" }
 maintenance = { status = "experimental" }
 
 [dependencies]
-chrono = "0.4"
-clap = "2.32"
-colored = "1.7"
-directories = "2.0"
-failure = "0.1"
-fern = { version = "0.5", features = ["colored"] }
 futures = "0.1"
-hex = "0.4"
-human-panic = { version = "1.0", optional = true }
-lazy_static = "1.2"
 log = "0.4"
-log-panics = { version = "2.0", features = ["with-backtrace"] }
-parking_lot = { version = "0.9", optional = true, features = ["deadlock_detection"] }
-rand = "0.7"
-serde = "1.0"
-serde_derive = "1.0"
 tokio = "0.1"
-toml = "0.5"
-url = "1.7"
 
-beserial = { path = "../beserial", version = "0.1" }
-nimiq-bls = { path = "../bls", version = "0.1" }
-nimiq-consensus = { path = "../consensus", version = "0.1" }
-nimiq-database = { path = "../database", version = "0.1" }
-nimiq-keys = { path = "../keys", version = "0.1" }
-nimiq-lib = { path = "../lib", version = "0.1" }
-nimiq-mempool = { path = "../mempool", version = "0.1" }
-nimiq-metrics-server = { path = "../metrics-server", version = "0.1", optional = true }
-nimiq-network = { path = "../network", version = "0.1" }
-nimiq-network-primitives = { path = "../network-primitives", version = "0.1" }
-nimiq-primitives = { path = "../primitives", version = "0.1", features = ["networks", "coin"] }
-nimiq-rpc-server = { path = "../rpc-server", version = "0.1", optional = true }
-nimiq-utils = { path = "../utils", version = "0.1" }
-
-[features]
-default = ["all"]
-#all = ["rpc-server", "metrics-server", "deadlock-detection", "human-panic"]
-all = ["rpc-server", "metrics-server", "deadlock-detection"]
-rpc-server = ["nimiq-rpc-server"]
-metrics-server = ["nimiq-metrics-server"]
-deadlock-detection = ["parking_lot"]
-system-install = []
+[dependencies.nimiq-lib]
+path = "../lib"
+version = "0.1"
+features = [
+    "validator",
+    "rpc-server",
+    "metrics-server",
+    "deadlock",
+    "logging",
+    "panic"
+]
