--- client/nimiq-client.service
@@ -1,15 +1,0 @@
-[Unit]
-Description=Nimiq's Rust client
-After=syslog.target network.target remote-fs.target nss-lookup.target
-
-[Service]
-Type=forking
-PIDFile=/run/nimiq-client.pid
-ExecStartPre=$(which nimiq-client)
-ExecStart=$(which nimiq-client)
-ExecReload=/bin/kill -s HUP $MAINPID
-ExecStop=/bin/kill -s QUIT $MAINPID
-PrivateTmp=true
-
-[Install]
-WantedBy=multi-user.target
