--- client/src/deadlock.rs
@@ -1,25 +1,0 @@
-use std::thread;
-use std::time::Duration;
-use parking_lot::deadlock;
-
-pub fn deadlock_detection() {
-    // Create a background thread which checks for deadlocks every 10s
-    thread::spawn(move || {
-        loop {
-            thread::sleep(Duration::from_secs(10));
-            let deadlocks = deadlock::check_deadlock();
-            if deadlocks.is_empty() {
-                continue;
-            }
-
-            error!("{} deadlocks detected", deadlocks.len());
-            for (i, threads) in deadlocks.iter().enumerate() {
-                error!("Deadlock #{}", i);
-                for t in threads {
-                    error!("Thread Id {:#?}", t.thread_id());
-                    error!("{:#?}", t.backtrace());
-                }
-            }
-        }
-    });
-}
