--- client2/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "nimiq-client2"
-version = "0.1.0"
-authors = ["The Nimiq Core Development Team <info@nimiq.com>"]
-edition = "2018"
-description = "Nimiq's Rust client"
-homepage = "https://nimiq.com"
-repository = "https://github.com/nimiq/core-rs-albatross"
-license = "Apache-2.0"
-categories = ["cryptography::cryptocurrencies"]
-keywords = ["nimiq", "cryptocurrency", "blockchain"]
-exclude = ["db", "peer_key.dat"]
-
-[badges]
-travis-ci = { repository = "nimiq/core-rs", branch = "master" }
-is-it-maintained-issue-resolution = { repository = "nimiq/core-rs" }
-is-it-maintained-open-issues = { repository = "nimiq/core-rs" }
-maintenance = { status = "experimental" }
-
-[dependencies]
-futures = "0.1"
-tokio = "0.1"
-log = "0.4"
-
-nimiq-lib2 = { path = "../lib2", version = "0.1", features = ["validator", "deadlock", "logging", "panic"] }
