+++ lib/Cargo.toml
@@ -16,22 +16,51 @@ is-it-maintained-issue-resolution = { repository = "nimiq/core-rs" }
 is-it-maintained-open-issues = { repository = "nimiq/core-rs" }
 maintenance = { status = "experimental" }
 
-[dependencies]
-failure = "0.1"
-futures = "0.1"
-lazy_static = "1.2"
-log = "0.4"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[dependencies]
+chrono = "0.4"
+colored = { version = "1.7", optional = true }
+derive_builder = "0.9"
+directories = "2.0"
+enum-display-derive = "0.1"
+failure = "0.1"
+fern = { version = "0.5", features = ["colored"], optional = true }
+hex = "0.4"
+human-panic = { version = "1.0", optional = true }
+lazy_static = "1.4"
+log = "0.4"
+log-panics = { version = "2.0", features = ["with-backtrace"], optional = true }
+parking_lot = { version = "0.7", features = ["deadlock_detection"], optional = true }
+paw = "1.0"
+rand = "0.7"
+rand04_compat = "0.1"
+serde = "1.0"
+serde_derive = "1.0"
+structopt = { version = "0.3", features = ["paw"] }
+toml = "0.5"
+url = "1.7"
+
+nimiq-blockchain-albatross = { path = "../blockchain-albatross", version = "0.1" }
 nimiq-bls = { path = "../bls", version = "0.1", optional = true }
 nimiq-consensus = { path = "../consensus", version = "0.1" }
 nimiq-database = { path = "../database", version = "0.1" }
+nimiq-keys = { path = "../keys", version = "0.1" }
 nimiq-mempool = { path = "../mempool", version = "0.1" }
+nimiq-metrics-server = { path = "../metrics-server", version = "0.1", optional = true }
 nimiq-network = { path = "../network", version = "0.1" }
 nimiq-network-primitives = { path = "../network-primitives", version = "0.1", features = ["all"] }
 nimiq-primitives = { path = "../primitives", version = "0.1", features = ["networks"] }
+nimiq-rpc-server = { path = "../rpc-server", version = "0.1", optional = true }
 nimiq-utils = { path = "../utils", version = "0.1" }
 nimiq-validator = { path = "../validator", version = "0.1", optional = true }
 
 [features]
-default = ["validator"]
-validator = ["nimiq-validator", "nimiq-bls"]
+default = []
+validator = ["nimiq-validator", "nimiq-bls", "nimiq-rpc-server/validator"]
+deadlock = ["parking_lot"]
+panic = ["log-panics", "human-panic"]
+logging = ["fern", "colored"]
+launcher = []
+rpc-server = ["nimiq-rpc-server"]
+metrics-server = ["nimiq-metrics-server"]
