+++ lib/src/config/consts.rs
@@ -1,3 +1,5 @@
+use std::net::{IpAddr, Ipv4Addr};
+
 
 /// The default port for `ws` and `wss`.
 pub const WS_DEFAULT_PORT: u16 = 8443;
@@ -8,8 +10,13 @@ pub const REVERSE_PROXY_DEFAULT_PORT: u16 = 8444;
 /// The default port for the RPC server
 pub const RPC_DEFAULT_PORT: u16 = 8648;
 
+/// The default port for the WebSocket server
+pub const WSRPC_DEFAULT_PORT: u16 = 8650;
+
 /// The default port for the metrics server
 pub const METRICS_DEFAULT_PORT: u16 = 8649;
 
-/// The default port for the WebSocket server
-pub const WEBSOCKET_DEFAULT_PORT: u16 = 8650;
+/// Returns the default bind, i.e. localhost
+pub fn default_bind() -> IpAddr {
+    IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1))
+}
