+++ lib/src/config/paths.rs
