+++ lib/src/error.rs
@@ -1,45 +1,118 @@
+use std::io::Error as IoError;
+
+use toml::de::Error as TomlError;
 use failure::Fail;
 
-use consensus::error::Error as ConsensusError;
+use database::lmdb::LmdbError;
+use database::volatile::VolatileDatabaseError;
 use network::error::Error as NetworkError;
+use utils::key_store::Error as KeyStoreError;
+use consensus::Error as ConsensusError;
+#[cfg(feature="validator")]
+use validator::error::Error as ValidatorError;
+#[cfg(feature="rpc-server")]
+use rpc_server::error::Error as RpcServerError;
 
 
-/// Prototype for a Error returned by these futures
-/// Errors can occur, when e.g. the bind port is already used
 #[derive(Debug, Fail)]
-pub enum ClientError {
-    #[fail(display = "{}", _0)]
-    NetworkError(#[cause] NetworkError),
-    #[fail(display = "Reverse Proxy can only be configured on Ws")]
-    ConfigureReverseProxyError,
-    #[fail(display = "{}", _0)]
-    ConsensusError(#[cause] ConsensusError),
-    #[fail(display = "Rtc is not implemented")]
-    RtcNotImplemented,
-    #[fail(display = "Protocol expects a hostname")]
-    MissingHostname,
-    #[fail(display = "Protocol expects a port")]
-    MissingPort,
-    #[fail(display = "Protocol doesn't expect a hostname")]
-    UnexpectedHostname,
-    #[fail(display = "Protocol doesn't expect a port")]
-    UnexpectedPort,
-    #[fail(display = "TLS identity file is missing")]
-    MissingIdentityFile,
-    #[fail(display = "Block producer error")]
-    BlockProducerError,
-    #[fail(display = "Client module failed")]
-    OtherFailed,
+pub enum Error {
+    #[fail(display = "Configuration error: {}", _0)]
+    Config(String),
+
+    #[fail(display = "LMDB error: {}", _0)]
+    Lmdb(#[cause] LmdbError),
+
+    #[fail(display = "I/O error: {}", _0)]
+    Io(#[cause] IoError),
+
+    #[fail(display = "Network error: {}", _0)]
+    Network(#[cause] NetworkError),
+
+    #[fail(display = "Key store error: {}", _0)]
+    KeyStore(#[cause] KeyStoreError),
+
+    #[fail(display = "Consensus error: {}", _0)]
+    Consensus(#[cause] ConsensusError),
+
+    #[fail(display = "Config file parsing error: {}", _0)]
+    Toml(#[cause] TomlError),
+
+    #[cfg(feature="validator")]
+    #[fail(display = "Validator error: {}", _0)]
+    Validator(#[cause] ValidatorError),
+
+    #[cfg(feature="rpc-server")]
+    #[fail(display = "RPC server error: {}", _0)]
+    RpcServer(#[cause] RpcServerError),
 }
 
-impl From<NetworkError> for ClientError {
-    fn from(e: NetworkError) -> Self {
-        ClientError::NetworkError(e)
+impl Error {
+    /// Constructs a configuration error from an error message.
+    ///
+    /// # Arguments
+    ///
+    /// * msg - The error message
+    ///
+    pub fn config_error<S: AsRef<str>>(msg: S) -> Self {
+        Self::Config(msg.as_ref().to_string())
     }
 }
 
-impl From<ConsensusError> for ClientError {
+impl From<LmdbError> for Error {
+    fn from(e: LmdbError) -> Self {
+        Self::Lmdb(e)
+    }
+}
+
+impl From<IoError> for Error {
+    fn from(e: IoError) -> Self {
+        Self::Io(e)
+    }
+}
+
+impl From<VolatileDatabaseError> for Error {
+    fn from(e: VolatileDatabaseError) -> Self {
+        match e {
+            VolatileDatabaseError::IoError(e) => e.into(),
+            VolatileDatabaseError::LmdbError(e) => e.into(),
+        }
+    }
+}
+
+impl From<NetworkError> for Error {
+    fn from(e: NetworkError) -> Self {
+        Self::Network(e)
+    }
+}
+
+impl From<KeyStoreError> for Error {
+    fn from(e: KeyStoreError) -> Self {
+        Self::KeyStore(e)
+    }
+}
+
+impl From<ConsensusError> for Error {
     fn from(e: ConsensusError) -> Self {
-        ClientError::ConsensusError(e)
+        Self::Consensus(e)
+    }
+}
+
+impl From<TomlError> for Error {
+    fn from(e: TomlError) -> Self {
+        Self::Toml(e)
+    }
+}
+
+#[cfg(feature="validator")]
+impl From<ValidatorError> for Error {
+    fn from(e: ValidatorError) -> Self {
+        Self::Validator(e)
+    }
+}
+
+#[cfg(feature="rpc-server")]
+impl From<RpcServerError> for Error {
+    fn from(e: RpcServerError) -> Self {
+        Self::RpcServer(e)
     }
 }
