+++ lib/src/extras/mod.rs
@@ -5,6 +5,10 @@ pub mod logging;
 pub mod deadlock;
 #[cfg(feature = "panic")]
 pub mod panic;
+#[cfg(feature = "rpc-server")]
+pub mod rpc_server;
+#[cfg(feature = "metrics-server")]
+pub mod metrics_server;
 
 #[cfg(feature = "launcher")]
 pub mod launcher;
