+++ lib/src/extras/panic.rs
