+++ lib/src/lib.rs
@@ -1,20 +1,40 @@
 #[macro_use]
+extern crate derive_builder;
+#[macro_use]
+extern crate failure;
+#[macro_use]
+extern crate enum_display_derive;
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate serde_derive;
+#[macro_use]
 extern crate lazy_static;
 
+extern crate nimiq_network as network;
 extern crate nimiq_consensus as consensus;
 extern crate nimiq_database as database;
-extern crate nimiq_network as network;
 extern crate nimiq_network_primitives as network_primitives;
 extern crate nimiq_primitives as primitives;
 extern crate nimiq_mempool as mempool;
 extern crate nimiq_utils as utils;
+extern crate nimiq_keys as keys;
+extern crate nimiq_blockchain_albatross as blockchain;
 
-#[cfg(feature = "validator")]
+#[cfg(feature="validator")]
 extern crate nimiq_validator as validator;
-#[cfg(feature = "validator")]
+#[cfg(feature="validator")]
 extern crate nimiq_bls as bls;
 
-pub mod prelude;
-pub mod client;
+#[cfg(feature="rpc-server")]
+extern crate nimiq_rpc_server as rpc_server;
+
+#[cfg(feature="metrics-server")]
+extern crate nimiq_metrics_server as metrics_server;
+
+
+pub mod config;
 pub mod error;
-pub mod block_producer;
+pub mod client;
+pub mod prelude;
+pub mod extras;
