+++ lib/src/prelude.rs
@@ -1,4 +1,8 @@
-pub use crate::client::{Client, ClientBuilder};
-pub use crate::error::ClientError;
-pub use database::Environment;
-pub use primitives::networks::NetworkId;
+
+#[cfg(feature = "validator")]
+pub use validator::validator::Validator;
+pub use crate::error::Error;
+pub use crate::config::config::ClientConfig;
+pub use crate::config::config_file::ConfigFile;
+pub use crate::config::command_line::CommandLine;
+pub use crate::client::{Client, Consensus};
