--- lib2/src/prelude.rs
@@ -1,8 +1,0 @@
-
-#[cfg(feature = "validator")]
-pub use validator::validator::Validator;
-pub use crate::error::Error;
-pub use crate::config::config::ClientConfig;
-pub use crate::config::config_file::ConfigFile;
-pub use crate::config::command_line::CommandLine;
-pub use crate::client::{Client, Consensus};
