+++ rpc-server/src/handlers/mod.rs
@@ -25,6 +25,7 @@ macro_rules! rpc_module_methods {
 
 pub mod consensus;
 pub mod block_production_nimiq;
+#[cfg(feature="validator")]
 pub mod block_production_albatross;
 pub mod blockchain;
 pub mod blockchain_nimiq;
@@ -34,6 +35,21 @@ pub mod mempool_albatross;
 pub mod network;
 pub mod wallet;
 
+
+pub use self::consensus::ConsensusHandler;
+pub use self::block_production_nimiq::BlockProductionNimiqHandler;
+#[cfg(feature="validator")]
+pub use self::block_production_albatross::BlockProductionAlbatrossHandler;
+pub use self::blockchain::BlockchainHandler;
+pub use self::blockchain_nimiq::BlockchainNimiqHandler;
+pub use self::blockchain_albatross::BlockchainAlbatrossHandler;
+pub use self::mempool::MempoolHandler;
+pub use self::mempool_albatross::MempoolAlbatrossHandler;
+pub use self::network::NetworkHandler;
+pub use self::wallet::{WalletHandler, UnlockedWalletManager};
+
+
 pub trait Module: Send + Sync {
     fn methods(self) -> Vec<(&'static str, Method)>;
 }
+
